+++ packages/dumb/src/components/stats/total-carbs/total-carbs-stat.css
@@ -34,7 +34,7 @@
   white-space: nowrap;
 }
 
-.rescueCarb, .mealCarb {
+.carbs{
   padding-left: 10px;
   margin-top: 4px;
 }
